pyproject.toml
[tool.poetry]
name = "downcomforter"
version = "0.1.0"
description = ""
authors = ["Jonathan Bowman <git@bowmanjd.com>"]

[tool.poetry.dependencies]
python = "^3.7"
livereload = "2.5.1"
"paka.cmark" = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
